requirements.txt
aiogram
supabase
requests
python-dotenv